Move output utils into their own file
